session och redirect in app och yardoc klart

+++ app.rb
@@ -31,36 +31,48 @@ before do
 end
 
 # Display Landing Page
-
-# @see Model#resetStart
 get('/') do
-    resetStart()
+    session[:loginError] = false
+    session[:registerError] = false
+    session[:like] = false
+    session[:empty] = false
+    session[:notUnique] = false
+    session[:isEmail] = true
+    session[:isNumber] = true
+    session[:auth] = false
 
     slim(:start)
 end
 
 # Displays a register form
-
-# @see Model#restartReg
 get('/showregister') do
-    restartReg()
+    session[:loginError] = false
+    session[:auth] = false
 
     slim(:register)
 end
 
 # Displays a login form
-
-# @see Model#restartLogin
 get('/showlogin') do
-    restartLogin()
+    session[:registerError] = false
+    session[:notUnique] = false
+    session[:isEmail] = true
+    session[:isNumber] = true
+    session[:auth] = false
     slim(:login)
 end
 
 # Displays Landing Page from logout
-
-# @see Model#resetStart
 get('/logout') do
-    resetStart()
+    session[:loginError] = false
+    session[:registerError] = false
+    session[:like] = false
+    session[:empty] = false
+    session[:notUnique] = false
+    session[:isEmail] = true
+    session[:isNumber] = true
+    session[:auth] = false
+
     slim(:start)
 end
 
@@ -68,23 +80,39 @@ end
 
 # @param [String] :filter, the filter to filter the posts by 
 
-# @see Model#restartPosts
 # @see Model#posts
 # @see Model#usernameAndId
-# @see Model#likeCountClient
+# @see Model#likeCountTotal
 # @see Model#likeCountPost
 # @see Model#isLiked
 get('/posts/:filter') do
-    restartPosts()
+    session[:notUnique] = false
+    session[:empty] = false
+    session[:isEmail] = true
+    session[:isNumber] = true
     filter = params[:filter]
 
-    posts = posts(filter)
-    usernameAndId = usernameAndId()
-    likeCountClient(session[:id])
-    likeCountPost = likeCountPost()
-    isLiked = isLiked()
+    postsAndFilter = posts(filter)
+
+    filterList = {
+        1 => "Woods",
+        2 => "Sea",
+        3 => "Mountains",
+        4 => "Lakes",
+        "all" => "All Posts"
+    }
+
+    filterId = postsAndFilter[:filterId]
+    session[:filter] = filterList[filterId]
+
     
-    slim(:"posts/index", locals:{posts:posts, usernameAndId:usernameAndId, isLiked:isLiked, likeCountPost:likeCountPost})
+
+    usernameAndId = usernameAndId()
+    session[:likeCount] = likeCountTotal(session[:id])
+    likeCountPost = likeCountPost()
+    isLiked = isLiked(session[:id])
+    
+    slim(:"posts/index", locals:{posts:postsAndFilter[:posts], usernameAndId:usernameAndId, isLiked:isLiked, likeCountPost:likeCountPost})
 end
 
 
@@ -93,11 +121,15 @@ end
 # @param [Integer] :userid, The client ID
 
 # @see Model#auth
-# @see Model#likeCountClient
+# @see Model#likeCountTotal
 get('/newpost/:userid') do
     userid = params[:userid].to_i
-    auth(userid)
-    likeCountClient(session[:id])
+
+    if not auth(session[:id], userid)
+        redirect("/error/401/")
+    end
+
+    session[:likeCount] = likeCountTotal(session[:id])
     slim(:"posts/new")
 end
 
@@ -107,24 +139,26 @@ end
 # @param [Integer] :postid, The post ID
 
 # @see Model#auth
-# @see Model#likeCountClient
+# @see Model#likeCountTotal
 # @see Model#post
 get('/post/:postid/:userid/edit') do
     userid = params[:userid].to_i
     postid = params[:postid].to_i
 
-    auth(userid)
-    likeCountClient(session[:id])
+    if not auth(session[:id], userid)
+        redirect("/error/401/")
+    end
+
+    session[:likeCount] = likeCountTotal(session[:id])
     post = post(postid)
 
     slim(:"posts/edit", locals:{post:post})
 end
 
-#Displays a single Profile
+# Displays a single Profile
 
 # @param [Integer] :userid, The client ID
 
-# @see Model#restartProfil
 # @see Model#users
 # @see Model#postSpecificInfo
 # @see Model#usersPersonality
@@ -132,19 +166,25 @@ end
 # @see Model#likeCountTotal
 # @see Model#likeCountPost
 # @see Model#isLiked
-# @see Model#likeCountClient
 get('/showprofile/:userid') do
     userid = params[:userid].to_i
 
-    restartProfil()
+    session[:loginError] = false
+    session[:registerError] = false
+    session[:like] = false
+    session[:empty] = false
+    session[:notUnique] = false
+    session[:isEmail] = true
+    session[:isNumber] = true
+
     userInfo = users(userid)
     postSpecificInfo = postSpecificInfo(userid)
     usersPersonality = usersPersonality(userid)
     usernameAndId = usernameAndId()
     likeCountTotal = likeCountTotal(userid)
     likeCountPost = likeCountPost()
-    isLiked = isLiked()
-    likeCountClient(session[:id])
+    isLiked = isLiked(session[:id])
+    session[:likeCount] = likeCountTotal(session[:id])
 
     slim(:"users/show", locals:{userInfo:userInfo, posts:postSpecificInfo, personality:usersPersonality, likesCount:likeCountTotal, usernameAndId:usernameAndId, likeCountPost:likeCountPost, isLiked:isLiked})
 end
@@ -154,13 +194,15 @@ end
 # @param [Integer] :userid, The client ID
 
 # @see Model#auth
-# @see Model#likeCountClient
+# @see Model#likeCountTotal
 # @see Model#users
 get('/user/:userid/edit') do
     userid = params[:userid].to_i
+    if not auth(session[:id], userid)
+        redirect("/error/401/")
+    end
 
-    auth(userid)
-    likeCountClient(session[:id])
+    session[:likeCount] = likeCountTotal(session[:id])
     user = users(userid).first
 
     slim(:"users/edit", locals:{user:user})
@@ -188,20 +230,86 @@ end
 
 # Attempts register 
 
+# @param [String] :email, The new user email
+# @param [String] :phonenumber, The new user phonenumber
+# @param [String] :password, The new user password
+# @param [String] :passwordConfirm, The new password confirm
+# @param [Integer] :username, The new user username
+# @param [Integer] :birthday, The new user birthday
+
 # @see Model#logTime
 # @see Model#emptyCredentials
 # @see Model#uniqueCredentials
-# @see Model#registration
+# @see Model#isEmail
+# @see Model#isNumber
+# @see Model#authenticationReg
+# @see Model#usersByUsername
+# @see Model#personalityUpdate
 post('/register') do
-    if logTime()
+    if session[:timeLogged] == nil
+        session[:timeLogged] = 0
+    end
+
+    logTime =  logTime(session[:timeLogged])
+
+    if logTime[:result]
+        session[:stress] = false
+        session[:timeLogged] = logTime[:time]
+
         credentials = [:username, :password, :passwordConfirm, :email, :phonenumber, :birthday]
         anyEmpty = emptyCredentials(credentials)
+
+        if anyEmpty
+            session[:empty] = true
+        else
+            session[:empty] = false
+        end
 
         credentials = [:username, :pwdigest, :email, :phonenumber]
         isNotUnique = uniqueCredentials(credentials)
 
-        registration(anyEmpty, isNotUnique)
+        if isNotUnique
+            session[:notUnique] = true
+        else
+            session[:notUnique] = false
+        end
+        
+        if not anyEmpty and not isNotUnique
+
+            if isEmail(params[:email])
+                session[:isEmail] = true
+            else
+                session[:isEmail] = false
+                redirect('/showregister')
+            end
+        
+            if isNumber(params[:phonenumber])
+                session[:isNumber] = true
+            else
+                session[:isNumber] = false
+                redirect('/showregister')
+            end
+
+            if authenticationReg(params[:password], params[:passwordConfirm], params[:username], params[:email], params[:phonenumber], params[:birthday])
+                user = usersByUsername(params[:username]).first
+                session[:id] = user["id"]
+                session[:user] = params[:username]
+                session[:auth] = true
+                session[:empty] = false
+                session[:notUnique] = false
+                personalityUpdate(session[:id])
+
+                redirect('/posts/all')
+            else
+                session[:registerError] = true
+                redirect('/showregister')
+            end
+        else
+            redirect('/showregister')
+        end  
     else
+        session[:stress] = true
+        session[:timeLogged] = logTime[:time]
         redirect('/showregister')
     end
 end
@@ -213,18 +321,45 @@ end
 
 # @see Model#logTime
 # @see Model#isEmpty
+# @see Model#usersByUsername
 # @see Model#authenticationLogin
 post('/login') do
-    if logTime()
+    if session[:timeLogged] == nil
+        session[:timeLogged] = 0
+    end
+
+    logTime =  logTime(session[:timeLogged])
+
+    if logTime[:result]
+        session[:stress] = false
+        session[:timeLogged] = logTime[:time]
+
         username = params[:username]
         password = params[:password]
 
         if isEmpty(username) || isEmpty(password)
+            session[:empty] = true
             redirect('/showlogin')
+        else
+            session[:empty] = false
         end
 
-        authenticationLogin(username, password)
+        user = usersByUsername(username).first
+
+        if authenticationLogin(username, password, user)
+            session[:loginError] = false
+            id = user["id"]
+            session[:id] = id
+            session[:auth] = true
+            session[:user] = username
+            redirect('/posts/all')
+        else
+            session[:loginError] = true
+            redirect('/showlogin')
+        end
     else
+        session[:stress] = true
+        session[:timeLogged] = logTime[:time]
         redirect('/showlogin')
     end
 end
@@ -239,26 +374,58 @@ end
 # @see Model#emptyCredentials
 # @see Model#isEmail
 # @see Model#isNumber
-# @see Model#upadteProfil
+# @see Model#updateProfile
 post('/user/:userid/update') do
     userid = params[:userid].to_i
-    if logTime()
+
+    if session[:timeLogged] == nil
+        session[:timeLogged] = 0
+    end
+
+    logTime =  logTime(session[:timeLogged])
+
+    if logTime[:result]
+        session[:stress] = false
+        session[:timeLogged] = logTime[:time]
+
         credentials = [:email, :phonenumber, :birthday]
         anyEmpty = emptyCredentials(credentials)
 
-        if not isEmail(params[:email])
+        if anyEmpty
+            session[:empty] = true
+        else
+            session[:empty] = false
+        end
+
+        if isEmail(params[:email])
+            session[:isEmail] = true
+        else
+            session[:isEmail] = false
             route = "/user/#{userid}/edit"
             redirect(route)
         end
 
-        if not isNumber(params[:phonenumber])
+        if isNumber(params[:phonenumber])
+            session[:isNumber] = true
+        else
+            session[:isNumber] = false
             route = "/user/#{userid}/edit"
             redirect(route)
         end
 
-        upadteProfil(userid, anyEmpty, credentials)
+        if updateProfile(userid, credentials)
+            session[:notUnique] = false
+            route = "/showprofile/#{userid}"
+            redirect(route)
+        else
+            session[:notUnique] = true
+            route = "/user/#{userid}/edit"
+            redirect(route)
+        end
 
     else
+        session[:stress] = true
+        session[:timeLogged] = logTime[:time]
         route = "/user/#{userid}/edit"
         redirect(route)
     end
@@ -270,15 +437,43 @@ end
 # @param [String] :text, The post text
 
 # @see Model#logTime
+# @see Model#isEmpty
 # @see Model#postNew
 post('/post/new') do
     id = session[:id]
-    if logTime()
+
+    if session[:timeLogged] == nil
+        session[:timeLogged] = 0
+    end
+
+    logTime =  logTime(session[:timeLogged])
+
+    if logTime[:result]
+        session[:stress] = false
+        session[:timeLogged] = logTime[:time]
+
         title = params[:title]
         text = params[:text]
 
-        postNew(title, text, id)
+        if not isEmpty(title)
+            session[:empty] = false
+
+            if postNew(title, text, id)
+                session[:notUnique] = false
+                redirect('/posts/all')
+            else
+                session[:notUnique] = true
+                route = "/newpost/#{id}"
+                redirect(route)
+            end
+        else
+            session[:empty] = true
+            route = "/newpost/#{id}"
+            redirect(route)
+        end
     else
+        session[:stress] = true
+        session[:timeLogged] = logTime[:time]
         route = "/newpost/#{id}"
         redirect(route)
     end
@@ -290,14 +485,43 @@ end
 # @param [String] :title, The new post title
 
 # @see Model#logTime
+# @see Model#isEmpty
 # @see Model#postUpdate
 post('/post/:id/update') do
     userid = session[:id]
     postid = params[:id].to_i
-    if logTime()
+
+    if session[:timeLogged] == nil
+        session[:timeLogged] = 0
+    end
+
+    logTime =  logTime(session[:timeLogged])
+
+    if logTime[:result]
+        session[:stress] = false
+        session[:timeLogged] = logTime[:time]
+
         title = params[:title]
-        postUpdate(title, postid, userid)
+
+        if not isEmpty(title)
+            session[:empty] = false
+
+            if postUpdate(title, postid, userid)
+                session[:notUnique] = false
+                redirect('/posts/all')
+            else
+                session[:notUnique] = true
+                route = "/post/#{postid}/#{userid}/edit"
+                redirect(route)
+            end
+        else
+            route = "/post/#{postid}/#{userid}/edit"
+            redirect(route)
+            session[:empty] = true
+        end 
     else
+        session[:stress] = true
+        session[:timeLogged] = logTime[:time]
         route = "/post/#{postid}/#{userid}/edit"
         redirect(route)
     end
@@ -308,12 +532,18 @@ end
 # @param [Integer] :userid, The client ID
 # @param [String] :postid, The post ID
 
+# @see Model#auth
 # @see Model#postDelete
 post('/post/:postid/:userid/delete') do
     userid = params[:userid].to_i
     postid = params[:postid].to_i
 
-    postDelete(userid, postid)
+    if not auth(session[:id], userid)
+        redirect("/error/401/")
+    end
+
+    postDelete(postid)
+    redirect('/posts/all')
 end
 
 # Likes a post
@@ -326,11 +556,35 @@ end
 # @see Model#filterRoute
 post('/post/:postid/:userid/like') do
     userid = params[:userid].to_i
-    auth(userid)
+    if not auth(session[:id], userid)
+        redirect("/error/401/")
+    end
 
     postid = params[:postid].to_i
     insertLike(userid, postid)
-    filterRoute()
+    filterId = filterRoute(session[:filter])
+
+    filterList = {
+        1 => "Woods",
+        2 => "Sea",
+        3 => "Mountains",
+        4 => "Lakes",
+        "all" => "All Posts"
+    }
+
+    filter = filterList[filterId]
+    session[:filter] = filter
+    if filterId == 1
+        redirect('/posts/woods')
+    elsif filterId == 2
+        redirect('/posts/sea')
+    elsif filterId == 3
+        redirect('/posts/mountains')
+    elsif filterId == 4
+        redirect('/posts/lakes')
+    elsif filterId == "all"
+        redirect('/posts/all')
+    end
 end
 
 # Unlikes a post
@@ -343,9 +597,33 @@ end
 # @see Model#filterRoute
 post('/post/:postid/:userid/unlike') do
     userid = params[:userid].to_i
-    auth(userid)
+    if not auth(session[:id], userid)
+        redirect("/error/401/")
+    end
 
     postid = params[:postid].to_i
     deleteLike(userid, postid)
-    filterRoute()
+    filterId = filterRoute(session[:filter])
+
+    filterList = {
+        1 => "Woods",
+        2 => "Sea",
+        3 => "Mountains",
+        4 => "Lakes",
+        "all" => "All Posts"
+    }
+
+    filter = filterList[filterId]
+    session[:filter] = filter
+    if filterId == 1
+        redirect('/posts/woods')
+    elsif filterId == 2
+        redirect('/posts/sea')
+    elsif filterId == 3
+        redirect('/posts/mountains')
+    elsif filterId == 4
+        redirect('/posts/lakes')
+    elsif filterId == "all"
+        redirect('/posts/all')
+    end
 end
